config feeds

+++ addons/Podcasts/tls.sh
@@ -17,8 +17,8 @@ delete=0
 play() {
 
     killall play
-    DCP="$DM_tl/Podcasts/.conf"
-    [ -f "$DCP/0.cfg" ] && st3=$(sed -n 2p "$DCP/0.cfg") || st3=FALSE
+    DIR2="$DM_tl/Podcasts/.conf"
+    [ -f "$DIR2/0.cfg" ] && st3=$(sed -n 2p "$DIR2/0.cfg") || st3=FALSE
     [ $st3 = FALSE ] && fs="" || fs='-fs'
     
     if [ -f "$DM_tl/Podcasts/cache/$2.mp3" ]; then
@@ -82,41 +82,40 @@ xmlns:atom='http://www.w3.org/2005/Atom'>
 </xsl:template>
 </xsl:stylesheet>"
 
-tmpl3="<?xml version='1.0' encoding='UTF-8'?>
-<xsl:stylesheet version='1.0'
-xmlns:xsl='http://www.w3.org/1999/XSL/Transform'
-xmlns:itunes='http://www.itunes.com/dtds/podcast-1.0.dtd'
-xmlns:media='http://search.yahoo.com/mrss/'
-xmlns:atom='http://www.w3.org/2005/Atom'>
-<xsl:output method='text'/>
-<xsl:template match='/'>
-<xsl:for-each select='/rss/channel/item'>
-<xsl:value-of select='enclosure/@url'/><xsl:text>-!-</xsl:text>
-<xsl:value-of select='media:cache[@type=\"image/jpeg\"]/@url'/><xsl:text>-!-</xsl:text>
-<xsl:value-of select='media:content[@type=\"image/jpeg\"]/@url'/><xsl:text>-!-</xsl:text>
-<xsl:value-of select='title'/><xsl:text>-!-</xsl:text>
-<xsl:value-of select='media:cache[@type=\"image/jpeg\"]/@duration'/><xsl:text>-!-</xsl:text>
-<xsl:value-of select='itunes:summary'/><xsl:text>-!-</xsl:text>
-<xsl:value-of select='description'/><xsl:text>EOL</xsl:text>
-</xsl:for-each>
-</xsl:template>
-</xsl:stylesheet>"
-    
+#tmpl3="<?xml version='1.0' encoding='UTF-8'?>
+#<xsl:stylesheet version='1.0'
+#xmlns:xsl='http://www.w3.org/1999/XSL/Transform'
+#xmlns:itunes='http://www.itunes.com/dtds/podcast-1.0.dtd'
+#xmlns:media='http://search.yahoo.com/mrss/'
+#xmlns:atom='http://www.w3.org/2005/Atom'>
+#<xsl:output method='text'/>
+#<xsl:template match='/'>
+#<xsl:for-each select='/rss/channel/item'>
+#<xsl:value-of select='enclosure/@url'/><xsl:text>-!-</xsl:text>
+#<xsl:value-of select='media:cache[@type=\"image/jpeg\"]/@url'/><xsl:text>-!-</xsl:text>
+#<xsl:value-of select='media:content[@type=\"image/jpeg\"]/@url'/><xsl:text>-!-</xsl:text>
+#<xsl:value-of select='title'/><xsl:text>-!-</xsl:text>
+#<xsl:value-of select='media:cache[@type=\"image/jpeg\"]/@duration'/><xsl:text>-!-</xsl:text>
+#<xsl:value-of select='itunes:summary'/><xsl:text>-!-</xsl:text>
+#<xsl:value-of select='description'/><xsl:text>EOL</xsl:text>
+#</xsl:for-each>
+#</xsl:template>
+#</xsl:stylesheet>"
     
     if [ -z "$2" ]; then
-    [ "$DCP/$3.rss" ] && rm "$DCP/$3.rss"; exit 1; fi
+    [ "$DIR2/$3.rss" ] && rm "$DIR2/$3.rss"; exit 1; fi
     feed="$2"
     num="$3"
-    DCP="$DM_tl/Podcasts/.conf"
+    DIR2="$DM_tl/Podcasts/.conf"
     xml="$(xsltproc - "$feed" <<< "$tmpl1" 2> /dev/null)"
     items1="$(echo "$xml" | tr '\n' ' ' | tr -s '[:space:]' \
     | sed 's/EOL/\n/g' | head -n 1 | sed -r 's|-\!-|\n|g')"
     xml="$(xsltproc - "$feed" <<< "$tmpl2" 2> /dev/null)"
     items2="$(echo "$xml" | tr '\n' ' ' | tr -s [:space:] \
     | sed 's/EOL/\n/g' | head -n 1 | sed -r 's|-\!-|\n|g')"
-    xml="$(xsltproc - "$feed" <<< "$tmpl3" 2> /dev/null)"
-    items3="$(echo "$xml" | tr '\n' ' ' | tr -s [:space:] \
-    | sed 's/EOL/\n/g' | head -n 1  | sed -r 's|-\!-|\n|g')"
+    #xml="$(xsltproc - "$feed" <<< "$tmpl3" 2> /dev/null)"
+    #items3="$(echo "$xml" | tr '\n' ' ' | tr -s [:space:] \
+    #| sed 's/EOL/\n/g' | head -n 1  | sed -r 's|-\!-|\n|g')"
 
     fchannel() {
         
@@ -136,51 +135,56 @@ xmlns:atom='http://www.w3.org/2005/Atom'>
             let n++
         done <<< "$items1"
     }
-   
+    
     ftype1() {
         
         n=1
         while read -r get; do
             [[ $n = 3 || $n = 5 || $n = 6 ]] && continue
-            if ([ -n "$(grep -o -E '\.mp3|\.mp4|\.ogg|\.avi|\.m4v|\.mov|\.flv' <<< "${get}")" ] && [ -z "$media" ]); then
+            if [ -n "$(grep -o -E '\.mp3|\.mp4|\.ogg|\.avi|\.m4v|\.mov|\.flv' <<< "${get}")" ] && [ -z "$media" ]; then
             media="$n"; type=1; break; fi
             let n++
         done <<< "$items2"
         
-        n=3
-        while read -r get; do
-            if ([ $(wc -w <<< "${get}") -ge 1 ] && [ $(wc -w <<< "${get}") -le 180 ] && [ -z "$title" ]); then
-            title="$n"; break; fi
-            let n++
-        done <<< "$items2"
+        f3="$(sed -n 3p <<<"$items2")"
+        f5="$(sed -n 5p <<<"$items2")"
+        f6="$(sed -n 6p <<<"$items2")"
+        
+        if [ $(wc -w <<< "${f3}") -ge 2 ] && [ $(wc -w <<< "${f3}") -le 200 ]; then
+        title=3; fi
 
-        n=5
-        while read -r get; do
-            if ([ $(wc -w <<< "${get}") -ge 1 ] && [ -z "$summ" ]); then
-            summ="$n"; break; fi
-            let n++
-        done <<< "$items2"
+        if [ $(wc -w <<< "${f5}") -ge 2 ] && [ -n "$(grep -o -E '\<|\>|/>' <<< "${f5}")" ]; then
+        sum1=5; fi
+        
+        if [ $(wc -w <<< "${f6}") -ge 2 ] && [ -n "$(grep -o -E '\<|\>|/>' <<< "${f6}")" ]; then
+        sum1=6; fi
+
+        if [ $(wc -w <<< "${f5}") -ge 2 ]; then
+        sum2=5; fi
+        
+        if [ $(wc -w <<< "${f6}") -ge 2 ]; then
+        sum2=6; fi
     }
     
     ftype2() {
 
         n=1
         while read -r get; do
-            if ([ -n "$(grep -o -E '\.jpg|\.jpeg|\.png' <<< "${get}")" ] && [ -z "$image" ]); then
+            if [ -n "$(grep -o -E '\.jpg|\.jpeg|\.png' <<< "${get}")" ] && [ -z "$image" ]; then
             image="$n"; type=2; break ; fi
             let n++
         done <<< "$items3"
         
         n=4
         while read -r get; do
-            if ([ $(wc -w <<< "${get}") -ge 1 ] && [ -z "$title" ]); then
+            if [ $(wc -w <<< "${get}") -ge 1 ] && [ -z "$title" ]; then
             title="$n"; break ; fi
             let n++
         done <<< "$items3"
         
         n=6
         while read -r get; do
-            if ([ $(wc -w <<< "${get}") -ge 1 ] && [ -z "$summ" ]); then
+            if [ $(wc -w <<< "${get}") -ge 1 ] && [ -z "$summ" ]; then
                 summ="$n"; break ; fi
             let n++
         done <<< "$items3"
@@ -190,13 +194,13 @@ xmlns:atom='http://www.w3.org/2005/Atom'>
 
         n=1
         while read -r get; do
-            if ([ -n "$(grep -E '\.jpg|\.jpeg|.png' <<< "${get}")" ] && [ -z "$image" ]); then
+            if [ -n "$(grep -E '\.jpg|\.jpeg|.png' <<< "${get}")" ] && [ -z "$image" ]; then
                 type=2
                 image="$n"; break; fi
-            if ([ -n "$(grep -o 'media:thumbnail url="[^"]*' | grep -o '[^"]*$')" <<< "${get}" ] && [ -z "$image" ]); then
+            if [ -n "$(grep -o 'media:thumbnail url="[^"]*' | grep -o '[^"]*$')" <<< "${get}" ] && [ -z "$image" ]; then
                 image="$n"; break; fi
                 type=2
-            if ([ -n "$(grep -o 'img src="[^"]*' | grep -o '[^"]*$')" <<< "${get}" ] && [ -z "$image" ]); then
+            if [ -n "$(grep -o 'img src="[^"]*' | grep -o '[^"]*$')" <<< "${get}" ] && [ -z "$image" ]; then
                 type=2
                 image="$n"; break; fi
             let n++
@@ -224,8 +228,12 @@ xmlns:atom='http://www.w3.org/2005/Atom'>
             get_summ
         fi
     fi
+    
+    if [ -z $sum2 ]; then
+    summary="$sum1"; else
+    summary="$sum2"; fi
 
-    if [[ -n "$title" && -n "$summ" && -z "$image" && -z "$media" ]]; then
+    if [[ -n "$title" && -n "$summary" && -z "$image" && -z "$media" ]]; then
     type=3; fi
     
     if [ "$type" = 1 ]; then
@@ -236,15 +244,16 @@ logo=\"$logo\"
 ntype=\"$type\"
 nmedia=\"$media\"
 ntitle=\"$title\"
-nsumm=\"$summ\"
+nsumm=\"$summary\"
 nimage=\"$image\"
 url=\"$feed\""
-        echo -e "$cfg" > "$DCP/$num.rss"; exit 0
+        echo -e "$cfg" > "$DIR2/$num.rss"; exit 0
         
     else
         url="$(tr '&' ' ' <<<"$feed")"
         msg "<b>$(gettext "Specified URL doesn't seem to contain any feeds")</b>\n$url\n\n<a href='leafpad'>\
 $(gettext "Tell us if you think this is an error")</a>" dialog-warning Idiomind &
+        [ "$DIR2/$num.rss" ] && "$DIR2/$num.rss"
         rm -f "$DT/cpt.lock"; exit 1
     fi
 }
@@ -252,19 +261,19 @@ $(gettext "Tell us if you think this is an error")</a>" dialog-warning Idiomind 
 check() {
 
     source $DS/ifs/mods/cmns.sh
-    DCP="$DM_tl/Podcasts/.conf"
+    DIR2="$DM_tl/Podcasts/.conf"
     DSP="$DS_a/Podcasts"
     [ -f "$DT/cpt.lock" ] && exit || touch "$DT/cpt.lock"
 
     internet
     
-    source "$DCP/$2.rss" # FIX
+    source "$DIR2/$2.rss" # FIX
 
     [ -z "$url" ] && exit 1
      
-    [ ! -f "$DCP/$2.rss" ] && printf "$tpl" > "$DCP/$2.rss"
+    [ ! "$DIR2/$2.rss" ] && printf "$tpl" > "$DIR2/$2.rss"
     
-    cp "$DCP/$2.rss" "$DCP/$2.rss_"
+    cp "$DIR2/$2.rss" "$DIR2/$2.rss_"
     
     podcast_items="$(xsltproc - "$url" <<<"$tmpl2" 2>/dev/null)"
     podcast_items="$(echo "$podcast_items" | tr '\n' ' ' | tr -s [:space:] | sed 's/EOL/\n/g' | head -n 2)"
@@ -272,8 +281,7 @@ check() {
     
     if [ -z "$(echo $item | sed 's/^ *//; s/ *$//; /^$/d')" ]; then
     msg "$(gettext "Couldn't download the specified URL")\n" info
-    rm -f "$DT/cpt.lock" & exit 1
-    fi
+    rm -f "$DT/cpt.lock" & exit 1; fi
     
     field="$(echo "$item" | sed -r 's|-\!-|\n|g')"
 
@@ -283,13 +291,13 @@ check() {
     --window-icon="$DS/images/icon.png" --columns=2 --skip-taskbar --scroll --on-top \
     --width=800 --height=600 --borders=5 \
     --text="\t<small>$(gettext "In this table you can define fields according to their content,  most of the time the default values is right. ")</small>" \
-    --field="":CB "$(sed -n 1p $DCP/$2.rss)!$mn" \
-    --field="":CB "$(sed -n 2p $DCP/$2.rss)!$mn" \
-    --field="":CB "$(sed -n 3p $DCP/$2.rss)!$mn" \
-    --field="":CB "$(sed -n 4p $DCP/$2.rss)!$mn" \
-    --field="":CB "$(sed -n 5p $DCP/$2.rss)!$mn" \
-    --field="":CB "$(sed -n 6p $DCP/$2.rss)!$mn" \
-    --field="":CB "$(sed -n 7p $DCP/$2.rss)!$mn" \
+    --field="":CB "$(sed -n 1p $DIR2/$2.rss)!$mn" \
+    --field="":CB "$(sed -n 2p $DIR2/$2.rss)!$mn" \
+    --field="":CB "$(sed -n 3p $DIR2/$2.rss)!$mn" \
+    --field="":CB "$(sed -n 4p $DIR2/$2.rss)!$mn" \
+    --field="":CB "$(sed -n 5p $DIR2/$2.rss)!$mn" \
+    --field="":CB "$(sed -n 6p $DIR2/$2.rss)!$mn" \
+    --field="":CB "$(sed -n 7p $DIR2/$2.rss)!$mn" \
     --field="":TXT "$(echo "$field" | sed -n 1p)" \
     --field="":TXT "$(echo "$field" | sed -n 2p)" \
     --field="":TXT "$(echo "$field" | sed -n 3p | sed 's/\://g')" \
@@ -299,46 +307,41 @@ check() {
     --field="":TXT "$(echo "$field" | sed -n 7p)" \
     --button=gtk-apply:0 | head -n 7 > $DT/f.tmp
 
-    [ -n "$(cat "$DT/f.tmp")" ] && mv -f $DT/f.tmp "$DCP/$2.rss" || cp -f "$DCP/$2.rss_" "$DCP/$2.rss"
-    [ -f "$DCP/$2.rss_" ] && rm "$DCP/$2.rss_"
+    [ -n "$(cat "$DT/f.tmp")" ] && mv -f $DT/f.tmp "$DIR2/$2.rss" || cp -f "$DIR2/$2.rss_" "$DIR2/$2.rss"
+    [ -f "$DIR2/$2.rss_" ] && rm "$DIR2/$2.rss_"
     [ -f "$DT/cpt.lock" ] && rm -f "$DT/cpt.lock" & exit
 }
 
 sync() {
    
-    DCP="$DM_tl/Podcasts/.conf"
+    DIR2="$DM_tl/Podcasts/.conf"
     cfg="$DM_tl/Podcasts/.conf/0.cfg"
     path="$(sed -n 3p < "$cfg" | grep -o 'path="[^"]*' | grep -o '[^"]*$')"
     
     if  [ -f "$DT/l_sync" ] && [ "$2" != A ]; then
-
-        msg_2 "$(gettext "A process is already running!")\n" info "OK" "gtk-stop" "Podcasts"
-        e=$(echo $?)
+    msg_2 "$(gettext "A process is already running!")\n" info "OK" "gtk-stop" "Podcasts"
+    e=$(echo $?)
         
-            if [ $e -eq 1 ]; then
-
-                killall rsync
-                [ -n "$(ps -A | pgrep -f "rsync")" ] && killall rsync
-                [ "$DT/cp.lock" ] && rm -f "$DT/cp.lock"
-                rm -f "$DT/l_sync"
-                killall tls.sh
-                exit 1
-            fi
+        if [ $e -eq 1 ]; then
+        killall rsync
+        [ -n "$(ps -A | pgrep -f "rsync")" ] && killall rsync
+        [ -f "$DT/cp.lock" ] && rm -f "$DT/cp.lock"
+        [ -f "$DT/cp.lock" ] && rm -f "$DT/l_sync"
+        killall tls.sh
+        exit 1; fi
             
-    elif  [ -f "$DT/l_sync" ] && [ "$2" = A ]; then
-    
-        exit 1
+    elif  [ -f "$DT/l_sync" ] && [ "$2" = A ]; then; exit 1
 
     elif [ ! -d "$path" ] && [ "$2" != A ]; then
-         
-        msg " $(gettext "The directory to synchronization does not exist \n Exiting.")" \
-        dialog-warning & exit 1
+    msg " $(gettext "The directory to synchronization does not exist \n Exiting.")" \
+    dialog-warning & 
+    [ -f "$DT/l_sync" ] && rm -f "$DT/l_sync"; exit 1
     
     elif  [ ! -d "$path" ] && [ "$2" = A ]; then
-            
-        echo "Synchronization error. Missing path" >> "$DM_tl/Podcasts/.conf/feed.err"
-        exit 1
-        
+    echo "Synchronization error. Missing path" >> "$DM_tl/Podcasts/.conf/feed.err"
+    [ -f "$DT/l_sync" ] && rm -f "$DT/l_sync"; exit 1
+    
+    
     elif [ -d "$path" ]; then
     
         set -e; set u
@@ -359,6 +362,7 @@ sync() {
         exit=$?
         
         if [ $exit = 0 ] ; then
+            [ -f "$DT/l_sync" ] && rm -f "$DT/l_sync"; exit 0
             st2="$(cd "$SYNCDIR"; ls *.mp3 | wc -l)"
             re=$((st2-st1))
                 if [ $re -gt 0 ]; then
@@ -366,12 +370,13 @@ sync() {
                 "$(gettext "Synchronization finished")" \
                 "$re $(gettext "New episode(s)")\n$st2 $(gettext "Total")" -t 8000; fi
         else
+            [ -f "$DT/l_sync" ] && rm -f "$DT/l_sync"; exit 1
             notify-send -i idiomind \
             "$(gettext "Error")" \
             "$(gettext "Error while syncing")" -t 8000
         fi
     fi
-    [ -f "$DT/l_sync" ] && rm -f "$DT/l_sync"; exit
+    
 }
 
 case "$1" in
